Move Mailing helper class to its own module.
Cleanup fix a typo.
Update history.
Add Mailing documentation to readme.
Cleanup update Mailing docstring.
