+++ docs/HISTORY.txt
@@ -5,7 +5,8 @@ Changelog
 1.4 (unreleased)
 ----------------
 
-- Nothing changed yet.
+- Move ``Mailing`` helper class to its own module ``mailing``.
+  [deif]
 
 
 1.3 (2013-05-03)
