+++ ftw/testing/pages.py
@@ -381,89 +381,3 @@ class PloneControlPanel(Plone):
         """
         return self.get_control_panel_links().get(title)
 
-
-class Mailing(object):
-    """The Mailing page object mocks the mailhost and captures sent emails.
-    The emails can then be easily used for assertions.
-
-    Usage:
-
-    >>> from ftw.testing.pages import Mailing
-    >>> import transaction
-
-    >>> class MyTest(TestCase):
-    ...     layer = MY_FUNCTIONAL_TESTING
-    ...
-    ...     def setUp(self):
-    ...         Mailing(self.layer['portal']).set_up()
-    ...         transaction.commit()
-    ...
-    ...     def tearDown(self):
-    ...         Mailing(self.layer['portal']).set_up()
-    ...
-    ...     def test_mail_stuff(self):
-    ...         portal = self.layer['portal']
-    ...         do_send_email()
-    ...         mail = Mailing(self.portal).pop()
-    ...         self.assertEquals('Subject: ...', mail)
-    """
-
-    def __init__(self, portal):
-        self.portal = portal
-
-    def set_up(self, configure=True):
-        """Setup a mock mail host so that emails can be catched and tested.
-        """
-
-        # Do NOT move the MockMailHost import to the top!
-        # It patches things implicitly!
-        from Products.CMFPlone.tests.utils import MockMailHost
-
-        mockmailhost = MockMailHost('MailHost')
-        self.portal.MailHost = mockmailhost
-        sm = self.portal.getSiteManager()
-        sm.registerUtility(component=mockmailhost,
-                           provided=IMailHost)
-
-        if configure:
-            mockmailhost.smtp_host = 'localhost'
-            self.portal.email_from_address = 'test@localhost'
-
-    def tear_down(self):
-        # Do NOT move the MockMailHost import to the top!
-        # It patches things implicitly!
-        from Products.CMFPlone.tests.utils import MockMailHost
-
-        sm = self.portal.getSiteManager()
-        mailhost = sm.getUtility(IMailHost)
-        if isinstance(mailhost, MockMailHost):
-            sm.unregisterUtility(component=mailhost, provided=IMailHost)
-
-    def get_mailhost(self):
-        mailhost = getUtility(IMailHost)
-
-        # Do NOT move the MockMailHost import to the top!
-        # It patches things implicitly!
-        from Products.CMFPlone.tests.utils import MockMailHost
-
-        assert isinstance(mailhost, MockMailHost), \
-            'The mailhost mocking was not set up properly. ' \
-            'Call ftw.testing.pages.Mailing().set_up() in your setUp method.'
-        return mailhost
-
-    def get_messages(self):
-        return self.get_mailhost().messages
-
-    def has_messages(self):
-        return len(self.get_messages()) > 0
-
-    def pop(self):
-        return self.get_messages().pop()
-
-    def reset(self):
-        # self.get_mailhost().reset()
-        try:
-            while self.pop():
-                pass
-        except IndexError:
-            pass
