+++ ftw/testing/tests/test_mailing.py
@@ -1,11 +1,11 @@
 from Products.MailHost.interfaces import IMailHost
-from ftw.testing.pages import Mailing
+from ftw.testing.mailing import Mailing
 from ftw.testing.testing import PAGE_OBJECT_FUNCTIONAL
 from unittest2 import TestCase
 from zope.component import getUtility
 
 
-class TestMailingPageObject(TestCase):
+class TestMailing(TestCase):
 
     layer = PAGE_OBJECT_FUNCTIONAL
